+++ python_modules/dagster/dagster_tests/core_tests/test_multiple_outputs.py
@@ -1,8 +1,10 @@
 import pytest
 
 from dagster import (
+    DagsterInstance,
     DagsterInvariantViolationError,
     DagsterStepOutputNotFoundError,
+    ExecutionTargetHandle,
     InputDefinition,
     Output,
     OutputDefinition,
@@ -73,7 +75,7 @@ def test_multiple_outputs_of_same_name_disallowed():
         execute_pipeline(muptiple_outputs_pipeline)
 
 
-def test_multiple_outputs_only_emit_one():
+def define_multi_out():
     @solid(
         name='multiple_outputs',
         input_defs=[],
@@ -85,12 +87,9 @@ def test_multiple_outputs_only_emit_one():
     def multiple_outputs(_):
         yield Output(output_name='output_one', value='foo')
 
-    called = {}
-
     @solid(name='downstream_one', input_defs=[InputDefinition('some_input')], output_defs=[])
     def downstream_one(_, some_input):
         del some_input
-        called['one'] = True
 
     @solid
     def downstream_two(_, some_input):
@@ -103,10 +102,48 @@ def test_multiple_outputs_only_emit_one():
         downstream_one(output_one)
         downstream_two(output_two)
 
-    result = execute_pipeline(multiple_outputs_only_emit_one_pipeline)
+    return multiple_outputs_only_emit_one_pipeline
+
+
+def test_multiple_outputs_only_emit_one():
+    result = execute_pipeline(define_multi_out())
     assert result.success
 
-    assert called['one']
+    solid_result = result.result_for_solid('multiple_outputs')
+    assert set(solid_result.output_values.keys()) == set(['output_one'])
+
+    with pytest.raises(
+        DagsterInvariantViolationError,
+        match="Output 'not_defined' not defined in solid 'multiple_outputs'",
+    ):
+        solid_result.output_value('not_defined')
+
+    with pytest.raises(DagsterInvariantViolationError, match='Did not find result output_two'):
+        solid_result.output_value('output_two')
+
+    with pytest.raises(
+        DagsterInvariantViolationError,
+        match=(
+            'Tried to get result for solid not_present in multiple_outputs_only_emit_one_pipeline. '
+            'No such top level solid.'
+        ),
+    ):
+        result.result_for_solid('not_present')
+
+    assert result.result_for_solid('downstream_two').skipped
+
+
+def test_multiple_outputs_only_emit_one_multiproc():
+    pipe = ExecutionTargetHandle.for_pipeline_python_file(
+        __file__, 'define_multi_out'
+    ).build_pipeline_definition()
+    result = execute_pipeline(
+        pipe,
+        environment_dict={'storage': {'filesystem': {}}, 'execution': {'multiprocess': {}}},
+        instance=DagsterInstance.local_temp(),
+    )
+    assert result.success
+
     solid_result = result.result_for_solid('multiple_outputs')
     assert set(solid_result.output_values.keys()) == set(['output_one'])