+++ CHANGES.md
@@ -21,6 +21,7 @@
   ```
   @composite_solid(config={}, config_fn=lambda config: {})
   ```
+
 - In the config system, `Dict` has been renamed to `Shape`; `List` to `Array`; `Optional` to `Noneable`; and `PermissiveDict` to `Permissive`. The motivation here is to clearly delineate config use cases versus cases where you are using types as the inputs and outputs of solids as well as python typing types (for mypy and friends). We believe this will be clearer to users in addition to simplifying our own implementation and internal abstractions.
 
   Our recommended fix is _not_ to used Shape and Array, but instead to use our new condensed config specification API. This allow one to use bare dictionaries instead of `Shape`, lists with one member instead of `Array`, bare types instead of `Field` with a single argument, and python primitive types (`int`, `bool` etc) instead of the dagster equivalents. These result in dramatically less verbose config specs in most cases.
@@ -43,6 +44,10 @@
   ```
 
   No imports and much simpler, cleaner syntax.
+
+**New**
+
+- `dagster/priority` tags can now be used to prioritize the order of execution for the built in in process and multiprocess engines.
 
 ## 0.6.7